+++ src/test/java/template/neodymium/tests/shadowdom/TextTest.java
@@ -1,4 +1,4 @@
-package template.neodymium.tests.smoke;
+package template.neodymium.tests.shadowdom;
 
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Selenide.$;
@@ -7,6 +7,7 @@ import static com.codeborne.selenide.Selenide.switchTo;
 import org.junit.Test;
 
 import com.codeborne.selenide.Selectors;
+import com.codeborne.selenide.Selenide;
 import com.codeborne.selenide.SelenideElement;
 
 import io.qameta.allure.Description;
@@ -17,7 +18,7 @@ import io.qameta.allure.junit4.DisplayName;
 import io.qameta.allure.junit4.Tag;
 import template.flows.OpenHomePageFlow;
 import template.neodymium.tests.AbstractTest;
-import template.pageobjects.pages.HomePage;
+import template.pageobjects.pages.ShadowDomPage;
 
 /**
  * @author kunze
@@ -28,19 +29,39 @@ import template.pageobjects.pages.HomePage;
 @DisplayName("TextTest")
 
 public class TextTest extends AbstractTest {
+    
+    @Test
+    @Description(value = "Check that Textfield exists and")
+    public void testVisitingPage() {
+        
+        // Goto the javadoc page
+        ShadowDomPage shadowdomPage = new ShadowDomPage();
+        Selenide.open("https://javascript.info/shadow-dom");
+
+        // short validation to check that the correct page was opened
+        shadowdomPage.isExpectedPage();
+        
+        // basic validation
+        shadowdomPage.validateStructure();
+        shadowdomPage.title.validateTitle("Shadow DOM");
+    }
 	
 	@Test
 	@Description(value = "Check that Textfield exists and")
 	public void checkTextExists() {
 		
-		//Open Homepage
-		HomePage homePage = OpenHomePageFlow.flow();
-		homePage.isExpectedPage();
+		//Open javadoc page
+	    ShadowDomPage shadowdomPage = new ShadowDomPage();
+	    Selenide.open("https://javascript.info/shadow-dom");
+	    
+	    // short validation to check that the correct page was opened
+	    shadowdomPage.isExpectedPage();
 		
 		//switch to iFrame of content
 		switchTo().frame($("iframe[style=\"height:60px\"]"));
 		
 		//Get the shadowDom-element that contains the text
+		//Syntax is shadowCss(element to be found, element that has the shadow Dom)
 		SelenideElement text = $(Selectors.shadowCss("p", "show-hello[name=John]"));
 		
 		//check if this exists
